rename to ciprs_reader
