+++ ciprs_reader/const.py
