+++ ciprs_reader/parser/lines.py
@@ -4,8 +4,8 @@ import datetime as dt
 import logging
 import re
 
-from ciprs.parser.base import Parser
-from ciprs.const import Section
+from ciprs_reader.parser.base import Parser
+from ciprs_reader.const import Section
 
 logger = logging.getLogger(__name__)
