+++ setup.py
@@ -1,5 +1,5 @@
 from setuptools import setup, find_packages
 
-import ciprs
+import ciprs_reader
 
-setup(name="ciprs_reader", version=ciprs.VERSION, packages=find_packages())
+setup(name="ciprs_reader", version=ciprs_reader.VERSION, packages=find_packages())
